Added login/register pages and react routing
