+++ frontend/src/App.css
@@ -257,4 +257,3 @@ footer {
   color: white;
   transform: scale(1.05);
 }
-
