+++ frontend/src/Footer.jsx
@@ -1,0 +1,12 @@
+const Footer = () => {
+    return(
+        <div>
+            {/* Footer Section */}
+            <footer>
+                <p>&copy; 2025 Your Company. All Rights Reserved.</p>
+            </footer>
+        </div>
+    )
+}
+
+export default Footer
